rastertiles

+++ rasterize.py
@@ -1,4 +1,4 @@
-import sys
+import os, sys
 import threading
 from queue import Queue
 import requests
@@ -18,10 +18,24 @@ def main():
         tiles = mbtiles.fetchtiles(tile_id)
         if len(tiles) < 1000:
             for tile in tiles:
-                queue.put(tilecoord(tile))
+                if tile[0] <= 10: # zoom
+                    queue.put(tilecoord(tile))
 
-        i += 1
-        if i == 100: break
+        #i += 1
+        #if i == 100: break
+
+    print queue.qsize()
+
+    download_all()
+
+
+def mainz():
+    mbtiles = Mbtiles(sys.argv[1])
+    print mbtiles.images()
+
+    for zoom in range(10, 11):
+        for row in mbtiles.gettilesiter(zoom):
+            queue.put(tilecoord(row))
 
     print queue.qsize()
 
@@ -36,7 +50,13 @@ def tilecoord(row):
 
 
 def download_tile(url, filename):
+    if os.path.exists(filename): return
+    
     res = requests.get(url)
+
+    if not os.path.exists(os.path.dirname(filename)):
+        os.makedirs(os.path.dirname(filename))
+
     with open(filename, "wb") as f:
         f.write(res.content)
 
@@ -47,11 +67,21 @@ def worker(num):
         tile = queue.get()
 
         print tile
+        url = 'http://localhost:8081/styles/terrain/%d/%d/%d.png' % (tile[2], tile[0], tile[1])
+        filename = 'tiles/%d/%d/%d.png' % (tile[2], tile[0], tile[1])
+
+        #download_tile(url, filename)
+        
+        url2 = 'http://localhost:8081/styles/terrain/%d/%d/%d@2x.png' % (tile[2], tile[0], tile[1])
+        filename2 = 'tiles@2x/%d/%d/%d.png' % (tile[2], tile[0], tile[1])
+
+        download_tile(url2, filename2)
 
         queue.task_done()
 
+
 def download_all():
-    num_worker = 4
+    num_worker = 1
     threads = []
 
     for i in range(num_worker):
@@ -62,6 +92,6 @@ def download_all():
     queue.join()
 
 if __name__ == '__main__':
-    main()
+    mainz()
